Caused SConstruct to only use -D__STDC_LIMIT_MACROS when <stdint.h> is included by <iostream>.
Move a temporary Sample structure in loader_xm.cpp to the outside of a for
loop. This changed the mysterious SIGSEGV that was tracing to the head of
a for loop into a mysterious SIGSEGV that traces into stl_vector.h. This
bug might be fixed by turning the vector<Sample> in this class into an
array, as it appears that vector is storing _references_ to the Sample
(which become invalid when the Sample goes out of scope), even though
the declaration fools us into expecting it to store the value.

Finally, exception handling has been added to the interface code that
supervises the module loading process. This means that if any of the
Loader_* module classes, or the functions they use, throw an exception,
it will be handled properly at the GUI level.

Sample loading still does not handle exceptions, however.


git-svn-id: https://svn.code.sf.net/p/cheesetracker/code@11 88f81a63-fe31-0410-a212-96c38ac4dcd5