Fixed the XM crash bug discovered in revision 19, but now the OSS driver will not allow itself to be enabled. No error message, of course.
The XM bug was caused by an Out_Of_Bounds exception that was being caught
in main(). Leaving the scope of QApplication::exec caused lots of stuff
in all threads to be deallocated, which lead to the crash, in a thread
other than the one main() was in.

This proves that exceptions must be caught within the Qt callback function
at the bottom of the call stack, before control is allowed to return to
Qt.

Also, exception handling has been added to all threads, and a fatal_error()
method has been added to the class Error, which prints the error message
and calls abort().



git-svn-id: https://svn.code.sf.net/p/cheesetracker/code@21 88f81a63-fe31-0410-a212-96c38ac4dcd5