Added proper error reporting to the OSS driver. The "bug" with the OSS driver described in r21 was simply an error that was going unreported.
A new module has been added: message_box.cpp. This is simply a wrapper
around QMessageBox::critical. It is needed because #including anything
from Qt in sound_driver_oss.cpp causes that file to fail to compile with
some really wierd errors.

It is now discovered that the resampler_* modules have extreme difficulty
staying within the boundaries of the samples they are given to mix. This
may be by design, but the original code accessed the data directly by
dereferencing an array. "Stupid" error handling was added, which simply
returns blank samples when invalid ones are requested, but the behavior
when the beginning of a sample is reached while incrementing backwards
was left alone, meaning that the program still aborts when playing a
pingpong loop sample.


git-svn-id: https://svn.code.sf.net/p/cheesetracker/code@23 88f81a63-fe31-0410-a212-96c38ac4dcd5