Lots and lots of changes here. I've added mutexes to the Sample_Data class, made minor modifications to the Sample_Data interface, fixed a few bugs, renamed some stupidly-named variables, and finally added an assert() into tracker_voice.cpp, which the resampler modules must pass.
About the assert(), it ensures that the expected number of samples were
actually extracted from the Sample_Data, in cases where this number can
be reliably calculated.

To pass the assert(), it is likely that the resampler subsystem needs
to be completely rewritten. It should be remembered that the mixing
buffer is passed to the existing resamplers through the "mixdata" member
"dst_buff" and the expected number of samples that should be stored in
this buffer is given in the "samples_to_mix" member.

I cannot confirm by reading the resampler code that there should be any
expectation that "samples_to_mix" is the actual number of samples placed
in the mixing buffer.


git-svn-id: https://svn.code.sf.net/p/cheesetracker/code@27 88f81a63-fe31-0410-a212-96c38ac4dcd5